update with upstream (and disable 0.6.3 mobile for now)

+++ source/mikolka/vslice/components/crash/Logger.hx
@@ -1,5 +1,6 @@
 package mikolka.vslice.components.crash;
 
+import haxe.Exception;
 import flixel.system.debug.log.LogStyle;
 import mikolka.compatibility.VsliceOptions;
 #if sys
@@ -9,23 +10,46 @@ import haxe.Log;
 
 class Logger{
     private static var file:FileOutput;
+    public static var enforceLogSettings:Bool = false;
     public static function startLogging() {
         #if LEGACY_PSYCH
         file = File.write("latest.log");
         #else
-        file = File.write(StorageUtil.getStorageDirectory()+"/latest.log");
+        try{
+            file = File.write(StorageUtil.getStorageDirectory()+"/latest.log");
+        }
+        catch(x:Exception){
+            #if (LEGACY_PSYCH)
+            FlxG.stage.window.alert(x.message, "File logging failed to init");
+            #else
+            CoolUtil.showPopUp(x.message,"File logging failed to init");
+            #end
+        }
         LogStyle.WARNING.onLog.add(log);
         LogStyle.ERROR.onLog.add(log);
         #end
         Log.trace = log;
     }
+    
     private static function log(v:Dynamic, ?infos:PosInfos):Void {
-        if(VsliceOptions.LOGGING == "None") return;
         var str = Log.formatOutput(v,infos);
-        if(VsliceOptions.LOGGING == "Console") Sys.println(str);
-        if(VsliceOptions.LOGGING == "File" && file != null) {
-            file.writeString(str+"\n");
-            file.flush();
+        if(enforceLogSettings){
+            if(VsliceOptions.LOGGING == "None") return;
+            if(VsliceOptions.LOGGING == "Console") Sys.println(str);
+            if(VsliceOptions.LOGGING == "File" && file != null) {
+                file.writeString(str+"\n");
+                file.flush();
+            }
+        }
+        else{
+            #if debug
+            Sys.println(str);
+            #else
+            if( file != null) {
+                file.writeString(str+"\n");
+                file.flush();
+            }
+            #end
         }
     }
 }
